feat: change HomeView View structure - Navigati\onLink
